Refactor REST API client package with request helpers
- Add `src/client/plantdb/client/rest_api/urls.py`
- Add `src/client/plantdb/client/rest_api/requests.py`
- Add `src/client/plantdb/client/rest_api/parsers.py`
